requirements.txt
streamlit
google-cloud-storage
google-cloud-texttospeech
google-cloud-aiplatform
pygame
python-dotenv